Model changes (inc Organization-Specific Elections)
This change allows us to attach individual elections to organizations.
This will let organizations only participate in elections they care
about. It also lets us better customize the app for the per-election
plans for each organization.

We also move to smalluuids for usernames, instead of the full long
UUIDs. This change will let us use shorter URLs when we start building
views around the UUID of individual account records. This migration
doesn't involve any data transformation because SmallUUID fields are
regular UUIDFields with extra sugar to make them look smaller.

Also included is the ability to assign a primary domain for an
organization and a boolean on the account to flag unsubscribed users.
This is in preparation of the initial bulk email functionality coming
shortly.