Prevent a migration, add editable=False
I changed it so that we inherit from the branding model mixin for the
organization_election model, which means the field `editable=False` was
added. Because Django migrations compare every arg and kwarg between
migrations django's `makemigrations` was causing a new migration to be
added just to add that non-database-modifying kwarg.

So if we add it to an older migration it won't cause a new migration.

This is a longer commit message than this change deserves.