Add django admin to make management easier
Only superusers will ever use the django admin, and there won't be a ton
of adding new records in the admin, but being able to edit the backend
elections and state records on the fly is going to be invaluable if/when
things change. We'll also have a 1000' view of users and be able to
promote admins of other organizations without going into the command
line.