Fix tests by allowing Organization's primary_domain to be null
To create an organization you have to create an Organization, then
create a domain attached to that organization, then edit the
Organization to have the primary domain be that organization. We'll need
to roll this into any management command or function we add to allow the
creation of new brandings. Not a priority right now though, but getting
tests working is!

This changes a migration to make a field nullable. Since neither staging
nor prod have run this migration (tests have been failing), I feel
comfortable making the change on the already-committed migration. Bad
form? Yes. Do I feel guilty about it? A little. Am I still doing it? Yes