Use separate 'high value' private S3 bucket for PII
With this change we'll be using a separate bucket specifically for 'high
value' assets such as uploads. This bucket will have private-by-default
permissions so that we're not exposing sensitive data.

We're also going to give prefixes to all uploaded non-static files in
order to reduce the chances of overwriting files, since there is not
generally a check to ensure unique-ness during S3 uploads.