Add the ability to send email via SES
Not 100% sure if we're using SES yet or something else, but for now
this adds a generic attachment we can use to send outgoing email. It
turns out that using the boto library directly makes sending email
more straightforward than trying to adapt some 3rd party python library
that doesn't do exactly what we want.

Added the config `APP_NAME` which will be passed through the ESP for
tracking and filtering later. So be sure to set that on various
deployments.