Crispy, Messages nav, jQuery, celery fix, dev uploads
We can now use crispy forms in the management interface, which means we
can use django forms that will have the attributes needed for bootstrap.

You can now use django's standard messaging framework in the management
interface for better user-friendly feedback.

Adding jQuery to the base template so that we can use bootstrap plugins
that are reliant on jQuery (i.e. all of them)

As we start to add background tasks changed it so that Celery tasks can
be set to run without a delay if needed, and if exceptions pop up in
those cases the exceptions bubble up to the top. Useful for debug and in
cases where we don't want to run a separate worker progress (yet.)

During development/debug /uploads/ will now serve files that have been
uploaded to the app.