Fix issue in branding middleware
Because we were using the same variable name (`domain`) in both the
cache key and as an object, we were accidentally passing that string to
the middleware. This fixes that.

Also, caching is broken on prod, because this should've showed up a long
time ago and only appeared now because I was setting up the staging
server which does have a valid cache running.