Add "Sending Address" (i.e. address emails come from)
This can be modified in the admin and selected by the organization. The
reason that this needs to be added in the admin is that you need to
whitelist the email address in SES and that process is going to be a
manual one for the time being.

Everyone will have access to the default sending address though, so even
if you don't have a whitelisted address you can still use the app.