Fix issue with primary_domain on Organization
Previously when creating an Organization you'd get an exception because
by default it wants to use the `primary_domain` with the pk of 0. That's
a problem mostly because no primary domain will ever have the primary
key of 0.

Basically instead of making the field False by default it should've been
None.