North Dakota doesn't register voters
It turns out North Dakota is unique in that it's the only state in the
nation that doesn't require anyone to register to vote. So we won't
ever send VR reminders to North Dakota.

From: https://vip.sos.nd.gov/pdfs/portals/votereg.pdf (4/23/18)
====
North Dakota is the only state without voter registration.
Although North Dakota was one of the first states to adopt
voter registration prior to the turn of the century, it abolished it
in 1951. It is also worth noting that North Dakota law still
provides cities with the ability to register voters for city
elections.

North Dakota is a rural state and its communities maintain
close ties and networks. North Dakota's system of voting, and
lack of voter registration, is rooted in its rural character by
providing small precincts. Establishing relatively small
precincts is intended to ensure that election boards know the
voters who come to the polling places to vote on Election Day
and can easily detect those who should not be voting in the
precinct.
====