For past and upcoming emails, use a OneToOneField
This is a pretty big refactor of how we've been doing things, so it
could use a bit of explanation.

Previously I was having identical fields on both records of past emails
(the "Mailing" model) and templates of potential upcoming emails (the
"MailingTemplate" model.) This made some sense, until you realized that
the app will, at times, treat these the same. The most obvious time it
will be treated the same is in the rendering of the emails, as there
may be times where we want to render old emails and times when we want
to render upcoming emails. You run into having to either duplicate the
rendering pipeline or having to put "isinstance" tests within the
rendering pipeline in order to render both types. This is pretty ugly,
and Django has an obvious fix for this with the OneToOneField. This
change breaks existing instances of the app and will require databases
to be refreshed, but that's a fine compromise if it saves us work in the
render pipeline.

Now I need to re-do the template creation and update pages in order to
create/update both the Email model and the Mailing model. Thankfully I
think I found code in Connect that will let me do this in a somewhat
straightforward way. Yay for reusing old code!