Allow VBM "Fault", "No Fault", and "All" targeting
Allow users to customize their messaging based on if the state requires
an excuse or not to request an absentee ballot. Some orgs may want to
only send to one type, others may want to send to everyone, others
may want to alter messaging based on if it requires an excuse or not.