Wrapper changes and optimizations
1) Make it impossible to un-default a default
You can change the default by creating or updating a non-default wrapper
to be the default. But you shouldn't be able to un-check the "default"
option on the default wrapper. It's too important that at least one
wrapper be the default.

2) Refactor template names
I built these templates before I really understood the template name
syntax for generic views. Renaming and consolidating these wrappers
keeps things clean and standardized across the app.