Render and preview emails and blocks 🇺🇸
We can now turn both emails and blocks into fully rendered and
customized HTML ready to be previewed to admins or sent to end-users.
Included here are new views to preview content, a function that sources
links, ways of linking to both the "Manage" and "Unsubscribe" pages
for individual users, and a bunch of complex queries and ORM goodness
which will find the exact block and target for each individual
recipient!

Oh, and we got the number of queries per email down to 5! And we don't
need to render individual blocks but can instead render the full body
of the email once! Less queries and less rendering than I anticipated!
Yay for code!

For the first time we're in a place where stakeholders can see the app
and the functionality. It's not pretty yet, but this is big!