Add 2018 data to the repo
Filling in the database with election data and final state data is
essentially the last big automated data part of the app. This data is
partially based off of publicly released data and verified internally by
searching Secretary of State websites. We'll want to keep an eye on this
as things change over the course of the year.

If you want to load the figure do
`./manage.py load data data/2018_election_data.json`