Streamline send calendar
It turns out `.annotate()` does nothing to `.raw()` queries (but raises
no exception or alert saying so) so we can remove that. And with
our election import we end up creating a bunch of pages really quick
so we need to both increase the number of items on each page and make it
so that our pagination doesn't list every possible page. This code
handles both instances.

Thanks to this StackOverflow thread that clarified what's possible with
RawQuerySets:
https://stackoverflow.com/questions/50122170/