Add index to EmailActivity to allow easier counting
We'll want an index we can use whenever we count activities from the
EmailActivity table. This is going to slow down the actual mailing
slightly so we'll want to keep an eye on this over time, and of course
significantly limit the number of indexes we have on that table.