Clarify the ordering of the send calendar
In the Send Calendar we didn't have a pre-set ordering for mailings
scheduled to go on the same date. Which was a problem mostly for testing
and since we want to have great test coverage and I want to keep my
sanity we're throwing an extra Template ID order on the calendar.