Restructure client assets and templates
Splitting out admin and constituent facing templates and assets to make
it easier to add to one without loading that into the other. This should
also make it easier to flesh out the admin further.

Also updated and streamlined the gulpfile. Previously we were building
on top of the gulpfile from a separate project. This is a new one just
for everyvoter.