Correctly set send_start, use task priorities for status updates
We weren't correctly marking the mailing as "sending" nor setting the
`send_start` field on the first send, so setting that. Also we didn't
prioritize status updates, which means some status updates that are
time-sensitive (like the ones that mark the start/end time, and those
that let admins know if sending is happening) may be delayed too long.