Exclude sent or sending messages from the calendar
This will prevent us from ever sending a template twice for the same
election, and also clean up the preview panel to no longer include
messages that have been sent that day or are pending being sent.

This will help us other places as well, but those are the two big ones.