Add election detail section
There is now an area for every election where you can see details about
that election like past and upcoming mailings, targeting, dates for
key deadlines, and more. Not fully fleshed out, but way more than we had
before and enough to get a good 10,000 foot view of where things are in
the platform.