No longer check for global unsubscribe before send
Hitting the database for a case-insensitive search for email is slowing
down sending. We already do checks for unsubscribe elsewhere in the app
so this should be unnecessary anyways.

In addition, if SES throws an error for an email being on the blacklist
we want to raise and capture that error instead of failing somewhat
silently. We'll have to address this soon.

We'll want to add global unsubscribe checks to the user import and email
sampling functionality.