Add paragraph breaks below blocks
There needs to be paragraph breaks below each block. If you want a break
before a block you'll need to add it to the actual top of the email. The
reason we don't automatically add a break between the top part of the
email and the first block is because it's possible that an email will
have no bottom part of the email and no blocks, which would cause an
unnecessary paragraph break pushing the bottom of the wrapper down.