Fix bug with template block widget
On the mailing template form there was a problem where if you clicked
"Edit Blocks" a modal would popup that would cause all the blocks to
be included in the template when the template is saved. To remove
blocks you'd then have to exit the template, re-enter the template, and
hit the "x" on the block without clicking "Edit Blocks"

By removing the <input> in the model, which I'm not 100% sure why it is
included in the first place, we fix this issue.